// NSDL_FVCI_UAT_RELEASE/13-03-2025/Website Deployment/nsdlfpi-development/src/swagger/model/draftFvciApplicationDto.ts
export interface DraftFvciApplicationDto { 
    id?: number;
    userId?: number;
    applicationId?: string;
    fvciRegistrationNumber?: string;
    createdAt?: Date;
    updatedAt?: Date;
    status?: number;
    kycDetails?: DraftFvicKycDetailsDto;
    kycLeiDetails?: Array<DraftFvciKycLeiDetailsDto>;
    addressDetails?: Array<DraftFvciAddressDetailsDto>;
    telephoneNumberDetails?: Array<DraftFvciTelephoneNumberDetailsDto>;
    contactDetails?: DraftFvciEkycContactDetailsDto;
    investmentManagerDetails?: Array<DraftFvciInvestmentManagerDetailsDto>;
    complianceOfficerDetails?: DraftFvciComplianceOfficerDetailsDto;
    complianceOfficerEmail?: DraftFvciComplianceOfficerEmailDto;
    kycDocuments?: Array<DraftFvciKycDocumentDto>;
    regulatoryAuthorityDetails?: DraftFvciRegulatoryAuthorityDetailsDto;
    globalCustodianDetails?: DraftFvciGlobalCustodianDetailsDto;
    declarationUndertakingDetails?: DraftFvciDeclarationUndertakingDetailsDto;
    isBank?: DraftFvciIsBankDto;
    kraPermission?: DraftFvciKraPermissionDto;
    ownershipDetails?: DraftFvciOwnershipDetailsDto;
    ekycBenificialOwnerDetails?: Array<DraftFvciEkycBenificialOwnerDetailsDto>;
    incomeDetails?: Array<DraftFvciIncomeDetailsDto>;
    registrationDetails?: DraftFvicRegistrationDetailsDto;
    benificialOwnershipByControlBoDetails?: Array<DraftFvciBenificialOwnershipByControlBoDetailsDto>;
    benificialOwnershipByControl?: DraftFvciBenificialOwnershipByControlDto;
    infoBasicOfOwnershipBoDetails?: Array<DraftFvciInfoBasicOfOwnershipBoDetailsDto>;
    informationOfSaSmFvciApplicant?: Array<DraftFvciInformationOfSaSmFvciApplicantDto>;
    subClassDetails?: Array<DraftFvciSubClassDetailsDto>;
    incidentsOfLawViolation?: DraftFvciIncidentsOfLawViolationDto;
    indianMarketAssocians?: DraftFvciIndianMarketAssociansDto;
}